refactor parser
